lantern-neo/requirements.txt
nextcord
python-dotenv
webcolors